refactor: Move MCP core modules to core directory
- Move MCP manager and utilities from utils/mcp to core/mcp
- Add error handling for invalid tool names
- Improve tool name display when server name is null
- Fix error handling in useToolCall function
