+++ src/core/mcp/exception.ts
@@ -1,0 +1,6 @@
+export class InvalidToolNameException extends Error {
+  constructor(name: string) {
+    super(`Invalid tool name: ${name}`)
+    this.name = 'InvalidToolNameException'
+  }
+}
