+++ src/core/mcp/tool-name-utils.ts
@@ -1,9 +1,12 @@
-const DEFAULT_DELIMITER = '__'
+import { InvalidToolNameException } from './exception'
+import { McpManager } from './mcpManager'
+
+const DEFAULT_DELIMITER = McpManager.TOOL_NAME_DELIMITER
 
 /**
  * Validates that a server name follows the required format and doesn't contain the delimiter
  * @param name Server name to validate
- * @param delimiter Optional custom delimiter (defaults to '---')
+ * @param delimiter Optional custom delimiter
  */
 export function validateServerName(
   name: string,
@@ -16,7 +19,6 @@ export function validateServerName(
       `Invalid MCP server name: ${name}. Only alphanumeric characters, underscores, and hyphens are allowed.`,
     )
   }
-  // Delimiter reserved for tool name construction (serverName---toolName)
   // Server names cannot contain it to ensure proper parsing and formatting
   if (name.includes(delimiter)) {
     throw new Error(
@@ -28,7 +30,7 @@ export function validateServerName(
 /**
  * Parses a combined tool name into server name and tool name components
  * @param name Combined tool name to parse
- * @param delimiter Optional custom delimiter (defaults to '---')
+ * @param delimiter Optional custom delimiter
  */
 export function parseToolName(
   name: string,
@@ -41,14 +43,14 @@ export function parseToolName(
   const match = name.match(regex)
 
   if (!match || match.length < 3) {
-    throw new Error(`Invalid tool name: ${name}`)
+    throw new InvalidToolNameException(name)
   }
 
   const serverName = match[1]
   const toolName = match[2]
 
   if (!serverName || !toolName) {
-    throw new Error(`Invalid tool name: ${name}`)
+    throw new InvalidToolNameException(name)
   }
 
   return { serverName, toolName }
@@ -58,7 +60,7 @@ export function parseToolName(
  * Creates a combined tool name from server name and tool name components
  * @param serverName Server name component
  * @param toolName Tool name component
- * @param delimiter Optional custom delimiter (defaults to '---')
+ * @param delimiter Optional custom delimiter
  */
 export function getToolName(
   serverName: string,
